rouge highlighter: refactor lexer and formatter selection (PR #3952)
As a followup to the discussion on
https://discuss.asciidoctor.org/Interpret-links-or-other-AsciiDoc-syntax-only-in-specific-lines-of-a-source-block-td8499.html#a8517
this patch moves the lexer and formatter selection of the Rouge
syntax highlighter into their own functions. This should make it easier
to extend the highlighter with custom lexers and/or highlighters without
having to replicate much of the logic in the highlight method.